pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mltoolbox"
version = "0.1.0"
description = "ML Development Environment Management Tools"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "python-dotenv>=0.19.0",
    "jinja2>=3.1.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
mltoolbox = "mltoolbox.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/mltoolbox"]

[tool.hatch.build.targets.wheel.force-include]
"src/mltoolbox/base" = "mltoolbox/base"
"src/mltoolbox/templates" = "mltoolbox/templates"